emmett changes
added some styling, changed movement to wasd, and started with some feature implementation. right now player health decreases when a wall is hit, but it is jank. the movement is also jank and the death state stopped working.  the movement does have some velocity functionality but something about it is weird, im sure you'll see.